Store each comment's type, and add PR review comments
It was feeling unwieldy to have separate lists of the different comment
types in a PullRequest. It felt like this was going to be especially
problematic for sorting comments. This commit refactors it so that there
is a single list of comments, and each comment knows its type. In
addition, it adds a comment type that was missing before: review
comments (i.e., the overall review).