Add handling of quote markers
GitHub treats a quoted tasklist as a tasklist (i.e., renders it as a
checkbox), so we'll treat it that way, too.

At first I thought that we'd generally not want to see quoted
checkboxes: they might arise accidentally when someone quotes a message
with checkboxes. But I can see circumstances where we *would* want to
see them. I'd rather err on the side of seeing more rather than less, so
I'll include these.