Add a --show-comments option
For now this is mutually exclusive with the --todo option, because it
seems messy to output both of them in succession. Eventually, I'm
thinking of adding an option to output to files in a directory rather
than stdout; in that case, you could provide both options, with each of
them going to a separate file.