Add a system test
When output isn't buffered (e.g., when the test fails), I get this
warning:

/Users/sacks/bin_external/github-tools/ghtools/gh_pr_query.py:31: ResourceWarning: unclosed <ssl.SSLSocket fd=6, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=0, laddr=('192.168.0.10', 51768), raddr=('140.82.114.5', 443)>
  access_token=access_token)
ResourceWarning: Enable tracemalloc to get the object allocation traceback

From a bit of searching, I see this reported in the python github api
repository. From some other reading, it sounds like this is just a
warning, and we can probably safely ignore it.