Change license to standard MIT license
The license file we were using before was based on the license in CESM
and related software. This was very similar to the MIT license but not
identical to it. I decided to keep things simple and just use the
standard MIT license - partly to support describing the license in
setup.py.